requirements.txt
aiohttp==3.7.4.post0
apturl==0.5.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
catfish==1.4.13
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==3.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
discord.py==1.7.0
distro==1.5.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
fasteners==0.14.1
future==0.18.2
gyp==0.1
httplib2==0.18.1
idna==2.10
jeepney==0.4.3
keyring==21.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.14.0
macaroonbakery==1.3.1
Magnus==1.0.3
Mako==1.1.2
MarkupSafe==1.1.1
menulibre==2.2.1
monotonic==1.5
multidict==5.1.0
netifaces==0.10.4
notify2==0.3
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
paramiko==2.7.1
pexpect==4.6.0
Pillow==7.2.0
protobuf==3.12.3
psutil==5.7.2
pycairo==1.16.2
pycodestyle==2.7.0
pycups==2.0.1
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
python-apt==2.1.3+ubuntu1.3
python-dateutil==2.8.1
python-debian==0.1.37
python-dotenv==0.17.0
python-xapp==2.0.1
pytz==2020.1
pyudev==0.21.0
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.47
requests==2.23.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==3.1.2
setproctitle==1.1.10
simplejson==3.17.0
six==1.15.0
systemd-python==234
toml==0.10.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
wadllib==1.3.4
xkit==0.0.0
yarl==1.6.3
youtube-dl==2020.9.14
